+++ Pandora/Paradigm/Controlflow/Joint.hs
@@ -2,6 +2,7 @@ module Pandora.Paradigm.Controlflow.Joint (module Exports) where
 
 import Pandora.Paradigm.Controlflow.Joint.Schemes as Exports
 import Pandora.Paradigm.Controlflow.Joint.Schematic as Exports
-import Pandora.Paradigm.Controlflow.Joint.Transformer as Exports
+import Pandora.Paradigm.Controlflow.Joint.Comonadic as Exports
+import Pandora.Paradigm.Controlflow.Joint.Monadic as Exports
 import Pandora.Paradigm.Controlflow.Joint.Interpreted as Exports
 import Pandora.Paradigm.Controlflow.Joint.Adaptable as Exports