+++ Pandora/Paradigm/Inventory/State.hs
@@ -5,7 +5,7 @@ module Pandora.Paradigm.Inventory.State (State (..), Stateful, current, modify, 
 import Pandora.Core.Functor (Variant (Co), type (:.), type (:=))
 import Pandora.Core.Morphism ((%))
 import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, unwrap))
-import Pandora.Paradigm.Controlflow.Joint.Transformer (Transformer (lay, wrap), (:>) (T))
+import Pandora.Paradigm.Controlflow.Joint.Monadic (Monadic (lay, wrap), (:>) (TM))
 import Pandora.Paradigm.Controlflow.Joint.Schematic (Schematic)
 import Pandora.Paradigm.Controlflow.Joint.Adaptable (Adaptable (adapt))
 import Pandora.Paradigm.Controlflow.Joint.Schemes.TUV (TUV (TUV))
@@ -56,9 +56,9 @@ instance Interpreted (State s) where
 
 type instance Schematic Monad (State s) u = TUV 'Co 'Co 'Co ((->) s) u ((:*:) s)
 
-instance Transformer (State s) where
-	lay x = T . TUV $ \s -> (s :*:) <$> x
-	wrap x = T . TUV $ point <$> unwrap x
+instance Monadic (State s) where
+	lay x = TM . TUV $ \s -> (s :*:) <$> x
+	wrap x = TM . TUV $ point <$> unwrap x
 
 type Stateful s = Adaptable (State s)