+++ CHANGELOG.md
@@ -145,3 +145,5 @@
 * Define `Category` type class with `identity` and `.` methods
 * Change `>->` signature: `a -> b` to `v a d` and `c -> d` to `v c d`
 * Extract `Schema` from `Transformer` type class as `Schematic` type family
+* Rename `Transformer` type class to `Monadic`
+* Define `Comonadic` type class for comonad transformers